double=>float。cuda で更新できるようにしてみた。しかし、途中で更新値がおかしくなる（浮動小数点数精度の問題？）

+++ DeepLearning_1_BP_Learning/data.h
@@ -4,5 +4,5 @@
 struct Instance
 {
 	unsigned int        cls;
-	std::vector<double> data;
+	std::vector<float> data;
 };
